Update ma901.component.ts
no need to use make a method async if it doesn't have await in it. Calling `await` on a non-promise can is not synchronous (this is complicated).
For example: 

a = await 1;

I think this will still add a new entry to system task queue, and other tasks already in the task queue will execute first. But I am not 100% sure.